test(integration): add full-stack integration tests for controllers
CategoryController: happy-path CRUD, 404 invalid-id, 400 bad payload

ExpenseController: create→list flow, 400 non-numeric amount

ReportController: monthly/weekly OK with valid X-API-KEY; 403 when key wrong or missing; MailSender & ReportService stubbed

StatisticsController: successful stats retrieval; 404 when user stats absent

UserController: create/read/update/picture/delete; duplicate-email & invalid-update negatives

CORS: allowed origin passes (GET + OPTIONS); disallowed origin blocked

Shared: assert async with asyncDispatch, use in-memory H2, isolate externals via @MockitoBean, no production code changes
