+++ src/main/resources/application-test.properties
@@ -1,0 +1,11 @@
+spring.datasource.url=jdbc:h2:mem:testdb;DB_CLOSE_DELAY=-1
+spring.datasource.driver-class-name=org.h2.Driver
+spring.datasource.username=sa
+spring.datasource.password=
+
+# JPA schema generation for tests
+spring.jpa.hibernate.ddl-auto=update
+spring.jpa.show-sql=false
+sendgrid.api.key=dummy-for-tests
+api.key=dummy-for-tests
+openai.secret=dummy-for-tests
